+++ CHANGELOG.md
@@ -16,9 +16,10 @@
 - CI-451 - Added new fields to International home page
 - CI-479 - Added video to Why choose UK Page
 - CI-480 - Added SEOOptimisationPanel to more pages
-
 - No ticket - Improve speed of test execution
-- 
+- XOT-1073 Add industry tags to Country Pages
+
+
 ## [2019.08.27_1](https://github.com/uktrade/directory-cms/releases/tag/2019.08.27_1)
 [Full Changelog](https://github.com/uktrade/directory-cms/compare/2019.08.27...2019.08.27_1)